refactor:region entity패키지 수정
