+++ src/main/java/com/sesacthon/foreco/category/entity/RegionCategory.java
@@ -3,7 +3,7 @@ package com.sesacthon.foreco.category.entity;
 import static jakarta.persistence.FetchType.*;
 
 import com.sesacthon.foreco.disposal.entity.Disposal;
-import com.sesacthon.foreco.disposal.controller.region.entity.Region;
+import com.sesacthon.foreco.region.entity.Region;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
