+++ src/main/java/com/sesacthon/foreco/member/service/MemberSignUpService.java
@@ -6,8 +6,8 @@ import com.sesacthon.foreco.member.dto.response.LoginResponseDto;
 import com.sesacthon.foreco.member.dto.response.MemberSimpleInfoResponse;
 import com.sesacthon.foreco.member.entity.Member;
 import com.sesacthon.foreco.member.entity.OAuth2Provider;
-import com.sesacthon.foreco.disposal.controller.region.entity.Region;
-import com.sesacthon.foreco.disposal.controller.region.service.RegionService;
+import com.sesacthon.foreco.region.entity.Region;
+import com.sesacthon.foreco.region.service.RegionService;
 import com.sesacthon.global.util.CookieUtil;
 import com.sesacthon.global.util.HttpHeaderUtil;
 import com.sesacthon.infra.feign.dto.response.KakaoUserInfoResponseDto;
