+++ src/main/java/com/sesacthon/foreco/region/repository/RegionRepository.java
@@ -1,6 +1,6 @@
-package com.sesacthon.foreco.disposal.controller.region.repository;
+package com.sesacthon.foreco.region.repository;
 
-import com.sesacthon.foreco.disposal.controller.region.entity.Region;
+import com.sesacthon.foreco.region.entity.Region;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
